pyproject.toml
[project]
name = "autobot-trader"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name", email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "numpy (>=2.2.6,<3.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "pyupbit (>=0.2.34,<0.3.0)",
    "ccxt (>=4.4.82,<5.0.0)",
    "python-telegram-bot (>=22.1,<23.0)",
    "schedule (>=1.2.2,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]
packages = [
    { include = "autobot_trader", from = "src" }
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
